Open the Java Newhall Graphical User Interface

This function must be called interactively.

newhall_GUI(command_only = FALSE)

Arguments

command_only

If TRUE return the command that would be executed to run GUI

Value

If intern=TRUE (default), the output of the command, one line per character string. 0 if successful. If the command could not be run for any reason, the value is 127 and a warning is issued.

Details

See documentation for system() return result for limitations on line length, error conditions, etc.