Run a Newhall simulation for each row in a CSV file

Run a Newhall simulation for each row in a CSV file

newhall_batch(
  .data = NULL,
  pathname = NULL,
  unitSystem = "metric",
  verbose = TRUE,
  toString = TRUE,
  checkargs = TRUE
)

Arguments

.data
a data.frame or character vector of paths (used as pathname)

pathname
path(s) to CSV file(s) with Newhall batch format and required columns; default NULL

unitSystem
Default: "metric" OR "mm" OR "cm" use millimeters of rainfall (default for the BASIC model); set to unitSystem="english" OR unitSystem="in" to transform English (inches of precipitation; degrees Fahrenheit) inputs to metric (millimeters of precipitation; degrees Celsius) before running simulation

verbose
print message about number of simulations and elapsed time

toString
call toString() method on each NewhallResults object and store in output column of result?

checkargs
logical; check argument length and data types for each run? Default: TRUE

Value

a data.frame containing list columns with Java Objects for NewhallDataset, NewhallResults. If toString=TRUE the column output is a character containing the toString() output from NewhallResults

See also