Load SoilWeb "ISSR-800" at 800 meter Resolution

Currently the only ISSR-800 data are only available for the contiguous (lower 48) United States. The only property cached for use in the Newhall model is the "available water holding capacity" (sum of storage for the whole profile). For consistency with PRISM grid the values are reprojected from "EPSG:5070" to "EPSG:4269".

newhall_issr800_subset(): Used to create a subset of the ISSR-800 soil available water storage data corresponding to the extent of an input spatial object x.

newhall_issr800_rast(): Create a SpatRaster object. This object contains Available Water Capacity (Storage) for at 800 meter resolution using the standard jNSM column naming scheme.

newhall_issr800_cache(
  ISSR800_PATH = file.path(newhall_data_dir("cache"), "SoilWeb", "800m"),
  template = newhall_nad83_template(),
  overwrite = FALSE
)

newhall_issr800_subset(
  x,
  template = newhall_nad83_template(),
  ISSR800_PATH = file.path(newhall_data_dir("cache"), "SoilWeb", "800m")
)

newhall_issr800_rast(
  ISSR800_PATH = file.path(newhall_data_dir("cache"), "SoilWeb", "800m"),
  tiffile = list.files(ISSR800_PATH, "\\.tif$", full.names = TRUE)
)

Arguments

ISSR800_PATH

Default: file.path(newhall_data_dir("cache"), "SoilWeb", "800m")

template

Template SpatRaster or target CRS specification for reprojection. Default: newhall_nad83_template()

overwrite

Force download of new cache files? Default: FALSE.

x

A SpatVector, SpatRaster, SpatExtent, or any other object type suitable to use with terra::crop();

tiffile

Optional: custom vector of paths to files to use to build raster. Defaults to all .TIF files in the specified cache directory and resolution.

Value

character. Path to cached water storage GeoTIFF.

References

Walkinshaw, Mike, A.T. O'Geen, D.E. Beaudette. "Soil Properties." California Soil Resource Lab, 1 Oct. 2022, https://casoilresource.lawr.ucdavis.edu/soil-properties/.

See also